* 'php-code-clean' of github.com:juizmill/ListService:
  remove duplication code in AbstractController
  edit documentor
  remove duplication code in AbstractController
  remove duplication code in AbstractController
  remove duplication code in AbstractController
  remove duplication code in AbstractController
  edit .scrutinizer.yml
  edit .scrutinizer.yml
  edit .scrutinizer.yml
  Created AbstractController, AbstractModel, AbstractForm

Conflicts:
	module/Application/src/Application/Controller/AbstractController.php
	module/Application/tests/ApplicationTest/Controller/AbstractControllerTest.php

+++ module/Application/src/Application/Model/AbstractModel.php
@@ -43,7 +43,7 @@ class AbstractModel implements ModelInterface
             $this->entityManager->persist($entityInterface);
         }
 
-        $this->entityManager->flush($entityInterface);
+        $this->entityManager->flush();
 
         return $entityInterface;
     }
@@ -53,10 +53,12 @@ class AbstractModel implements ModelInterface
      */
     public function remove($identity)
     {
-        $entity = $this->getReference($identity);
+        $entity = $this->getRepository()->find((int)$identity);
 
-        $this->entityManager->remove($entity);
-        $this->entityManager->flush($entity);
+        if ($entity) {
+            $this->entityManager->remove($entity);
+            $this->entityManager->flush();
+        }
 
         return $entity;
     }
